+++ taskmanager_api/admin.py
@@ -1,15 +1,7 @@
 """ Task Manager Admin"""
 from django.contrib import admin
-from django.contrib.admin.utils import flatten_fieldsets
-from django.contrib.auth.admin import UserAdmin
 
-from taskmanager_api.models import Comment, EventLog, User, RejectedTask, Task
-
-UserAdmin.fieldsets += (
-    ('Other Information',
-     {'fields': ('profile_image',)}
-     ),
-)
+from taskmanager_api.models import Comment, EventLog, RejectedTask, Task
 
 
 class TaskAdmin(admin.ModelAdmin):
@@ -26,6 +18,5 @@ class TaskAdmin(admin.ModelAdmin):
 
 admin.site.register(Comment)
 admin.site.register(EventLog)
-admin.site.register(User, UserAdmin)
 admin.site.register(RejectedTask)
 admin.site.register(Task, TaskAdmin)
